Early out optimization for isVisible
When there are many ractive viewport elements on a page, the recursive
nature of isInParents() leads to a lot of CPU usage even when these
viewports are not being displayed. This eliminates that usage when the
element is not visible, which can be detected through the offsetParent
property in modern browsers. If this property is not available,
isVisible() will continue to function as before.